backend/requirements.txt
fastapi
uvicorn
pydantic[email]
google-generativeai
python-dotenv
motor
python-jose[cryptography]
passlib[bcrypt]
bcrypt
httpx
python-multipart
requests
